+++ StripeCryptoOnramp/StripeCryptoOnramp/Source/Internal/API Bindings/STPAPIClient+CryptoOnramp.swift
@@ -63,7 +63,7 @@ extension STPAPIClient {
         let endpoint = "crypto/internal/kyc_requirements"
         return try await get(
             resource: endpoint,
-            consumerSessionClientSecret: consumerSessionClientSecret
+            additionalHeaders: [CryptoOnrampAPI.consumerAuthTokenHeader: consumerSessionClientSecret]
         )
     }
 
@@ -153,7 +153,7 @@ extension STPAPIClient {
         let endpoint = "crypto/internal/identifier_requirements"
         return try await get(
             resource: endpoint,
-            consumerSessionClientSecret: consumerSessionClientSecret
+            additionalHeaders: [CryptoOnrampAPI.consumerAuthTokenHeader: consumerSessionClientSecret]
         )
     }
 
@@ -198,7 +198,7 @@ extension STPAPIClient {
         let endpoint = "crypto/internal/crs_carf_declaration"
         return try await get(
             resource: endpoint,
-            consumerSessionClientSecret: consumerSessionClientSecret
+            additionalHeaders: [CryptoOnrampAPI.consumerAuthTokenHeader: consumerSessionClientSecret]
         )
     }
 
@@ -218,6 +218,53 @@ extension STPAPIClient {
         let endpoint = "crypto/internal/crs_carf_declaration"
         let requestObject = EmptyRequestWithCredentials(consumerSessionClientSecret: consumerSessionClientSecret)
         return try await post(resource: endpoint, object: requestObject)
+    }
+
+    /// Retrieves the current partner declaration state for the current Link user.
+    /// - Parameters:
+    ///   - declarationType: The type of declaration to retrieve.
+    ///   - linkAccountInfo: Information associated with the Link account including the client secret and whether the account has been verified.
+    /// - Returns: The current partner declaration state.
+    /// Throws if the `linkAccountSessionState` is not verified, a client secret doesn’t exist, or if an API error occurs.
+    func retrievePartnerTerms(declarationType: PartnerDeclarationType, linkAccountInfo: PaymentSheetLinkAccountInfoProtocol) async throws -> PartnerTerms {
+        guard let consumerSessionClientSecret = linkAccountInfo.consumerSessionClientSecret else {
+            throw CryptoOnrampAPIError.missingConsumerSessionClientSecret
+        }
+
+        try validateSessionState(using: linkAccountInfo)
+
+        let endpoint = "crypto/internal/partner_terms"
+        let request = RetrievePartnerTermsRequest(declarationType: declarationType)
+        return try await get(
+            resource: endpoint,
+            parameters: try request.encodeJSONDictionary(),
+            additionalHeaders: [CryptoOnrampAPI.consumerAuthTokenHeader: consumerSessionClientSecret]
+        )
+    }
+
+    /// Confirms the current Link user accepted a partner declaration.
+    /// - Parameters:
+    ///   - declarationId: The unique identifier of the declaration accepted by the customer.
+    ///   - linkAccountInfo: Information associated with the Link account including the client secret and whether the account has been verified.
+    /// - Returns: An empty response.
+    /// Throws if the `linkAccountSessionState` is not verified, a client secret doesn’t exist, or if an API error occurs.
+    @discardableResult
+    func confirmPartnerTerms(declarationId: String, linkAccountInfo: PaymentSheetLinkAccountInfoProtocol) async throws -> EmptyResponse {
+        guard let consumerSessionClientSecret = linkAccountInfo.consumerSessionClientSecret else {
+            throw CryptoOnrampAPIError.missingConsumerSessionClientSecret
+        }
+
+        try validateSessionState(using: linkAccountInfo)
+
+        let endpoint = "crypto/internal/partner_terms"
+        let requestObject = ConfirmPartnerTermsRequest(
+            declarationId: declarationId
+        )
+        return try await post(
+            resource: endpoint,
+            object: requestObject,
+            additionalHeaders: [CryptoOnrampAPI.consumerAuthTokenHeader: consumerSessionClientSecret]
+        )
     }
 
     /// Begins an identity verification session, providing the necessary data used to initialize the Identity SDK.
@@ -400,13 +447,19 @@ extension STPAPIClient {
 }
 
 private extension STPAPIClient {
+
     /// Helper method to wrap the closure-based post method for Swift concurrency.
-    func post<T: Decodable>(resource: String, object: Encodable) async throws -> T {
+    func post<T: Decodable>(
+        resource: String,
+        object: Encodable,
+        additionalHeaders: [String: String] = [:]
+    ) async throws -> T {
         return try await withCheckedThrowingContinuation { continuation in
             post(
                 resource: resource,
                 object: object,
-                apiVersionOverride: CryptoOnrampAPI.stripeAPIVersion
+                apiVersionOverride: CryptoOnrampAPI.stripeAPIVersion,
+                additionalHeaders: additionalHeaders
             ) { (result: Result<T, Error>) in
                 continuation.resume(with: result)
             }
@@ -414,13 +467,18 @@ private extension STPAPIClient {
     }
 
     /// Helper method to wrap the closure-based delete method for Swift concurrency.
-    func delete<T: Decodable>(resource: String, object: Encodable) async throws -> T {
+    func delete<T: Decodable>(
+        resource: String,
+        object: Encodable,
+        additionalHeaders: [String: String] = [:]
+    ) async throws -> T {
         let parameters = try object.encodeJSONDictionary()
         return try await withCheckedThrowingContinuation { continuation in
             delete(
                 resource: resource,
                 parameters: parameters,
-                apiVersionOverride: CryptoOnrampAPI.stripeAPIVersion
+                apiVersionOverride: CryptoOnrampAPI.stripeAPIVersion,
+                additionalHeaders: additionalHeaders
             ) { (result: Result<T, Error>) in
                 continuation.resume(with: result)
             }
@@ -428,49 +486,20 @@ private extension STPAPIClient {
     }
 
     /// Helper method to wrap the closure-based get method for Swift concurrency.
-    func get<T: Decodable>(resource: String, parameters: [String: Any] = [:]) async throws -> T {
+    func get<T: Decodable>(
+        resource: String,
+        parameters: [String: Any] = [:],
+        additionalHeaders: [String: String] = [:]
+    ) async throws -> T {
         return try await withCheckedThrowingContinuation { continuation in
             get(
                 resource: resource,
                 parameters: parameters,
-                apiVersionOverride: CryptoOnrampAPI.stripeAPIVersion
+                apiVersionOverride: CryptoOnrampAPI.stripeAPIVersion,
+                additionalHeaders: additionalHeaders
             ) { (result: Result<T, Error>) in
                 continuation.resume(with: result)
             }
-        }
-    }
-
-    func get<T: Decodable>(
-        resource: String,
-        parameters: [String: Any] = [:],
-        consumerSessionClientSecret: String
-    ) async throws -> T {
-        let url = apiURL.appendingPathComponent(resource)
-        var request = configuredRequest(
-            for: url,
-            apiVersionOverride: CryptoOnrampAPI.stripeAPIVersion,
-            additionalHeaders: [
-                CryptoOnrampAPI.consumerAuthTokenHeader: consumerSessionClientSecret,
-            ]
-        )
-        request.stp_addParameters(toURL: parameters)
-        request.httpMethod = "GET"
-
-        return try await withCheckedThrowingContinuation { continuation in
-            urlSession.stp_performDataTask(
-                with: request,
-                completionHandler: { data, response, error in
-                    DispatchQueue.main.async {
-                        let result: Result<T, Error> = STPAPIClient.decodeResponse(
-                            data: data,
-                            error: error,
-                            response: response,
-                            request: request
-                        )
-                        continuation.resume(with: result)
-                    }
-                }
-            )
         }
     }
 }